# Make compilation faster and less demanding