@svileng 2020-12-06
This is now published on hex as `1.2.0` 🎉